READ
My first design is a custom Datsun truck Hotwheel that is being wheel swapped. I chose this project because its something that I did growing up was swapping wheels on My Hotwheels, so this project was a super fun one! My first decision was on which wheels to use,  and ultimately went with the five spoke star style wheels. The first struggle I had was finding the correct shape of the star in the center of the wheel. I started with just drawing the center hole and triangle and crated a pattern according to the dimensions. The body of the truck has allot of curves, and with that brought struggles with drawing surfaces, but with the new plane that I can reference from an existing plane was the only way I was ale to get the holes for the wide body kit. Which was I big debate I had whither the truck needed the extra flairs, another detail I chose was making the truck right hand drive.  At first it's different to think of everything being chipped away from a block but understanding that helped allot with how I drew up some of the parts.
Starting from my sketches I felt far from the end, but as I went along and started getting farther into the project I grew confident. I hope you enjoy the hard work I put into my design for the different wheel set up. Each part had unique challenges, but with each change I could see the end picture of what I wanted in my head. Lining everything back up in the assembly was probably the most tedious part. As was adding pieces together, I thought one part was measured correctly the assebly said otherwise, non the less with small changes I was able to recreate the design with the new wheels.